requirements.txt
# local package
-e .

# external requirements
click~=8.1.3
Sphinx
coverage
awscli == 1.27.38
flake8
python-dotenv>=0.5.1

# backwards compatibility
pathlib2

# basics
matplotlib==3.6.2
numpy==1.24.1
pandas==1.5.2
tqdm==4.64.1

# project stuff
diffusers[training]==0.5.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
hydra-core==1.3.1
wandb==0.13.7
dvc==2.38.1
dvc[gs]
click==8.1.3
tensorboard==2.11.0
snakeviz==2.1.1
pytorch_lightning==1.8.6
google-api-python-client
PyYAML==5.4